# www/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "task.api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "TaskService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/task/health-check": {
      "get": {
        "operationId": "TaskService_HealthCheck",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1HealthCheckResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "TaskService"
        ]
      }
    },
    "/api/task/{taskID}": {
      "put": {
        "operationId": "TaskService_SetTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SetTaskResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "taskID",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "action": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "TaskService"
        ]
      }
    }
  },
  "definitions": {
    "Taskstatuses": {
      "type": "string",
      "enum": [
        "Pending",
        "Draft",
        "Returned",
        "Approved",
        "Rejected"
      ],
      "default": "Pending"
    },
    "Tasksteps": {
      "type": "string",
      "enum": [
        "Maker",
        "Checker",
        "Signer",
        "Releaser"
      ],
      "default": "Maker"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1HealthCheckResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "v1SaveTaskResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "v1SetTaskResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "boolean"
        },
        "code": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/v1Task"
        }
      }
    },
    "v1Task": {
      "type": "object",
      "properties": {
        "TaskID": {
          "type": "string",
          "format": "uint64",
          "readOnly": true
        },
        "type": {
          "type": "string",
          "required": [
            "type"
          ]
        },
        "status": {
          "$ref": "#/definitions/Taskstatuses"
        },
        "step": {
          "$ref": "#/definitions/Tasksteps"
        },
        "createdByID": {
          "type": "string",
          "format": "uint64",
          "required": [
            "createdByID"
          ]
        },
        "lastApprovedByID": {
          "type": "string",
          "format": "uint64"
        },
        "lastRejectedByID": {
          "type": "string",
          "format": "uint64"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "readOnly": true
        }
      },
      "required": [
        "type",
        "createdByID"
      ]
    }
  }
}